// HelpDeskFE/HelpDeskFETp/src/api/model/createTypeMaintenanceDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateTypeMaintenanceDto {
  /**
     * @minLength 0
     * @maxLength 60
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 10080
     */
  estimatedTime: number;
}