// HelpDeskFE/HelpDeskFETp/src/api/model/updateTypeMaintenanceDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface UpdateTypeMaintenanceDto {
  /**
     * @minLength 0
     * @maxLength 60
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 10080
     */
  estimatedTime: number;
}